Added buzz on passive speaker

+++ src/altimeter/snd.h
@@ -7,7 +7,7 @@ void termSound();
 void noSound();
 
 #if defined(SOUND_PASSIVE) || defined(SOUND_ACTIVE) || defined(SOUND_EXTERNAL)
-void sound(uint16_t frequency, uint8_t duration);
+void sound(uint16_t frequency, uint16_t duration);
 #else
 #define sound(a,b)
 #endif
